split GuiContext
