+++ src/main/java/com/cleanroommc/modularui/api/IMuiScreen.java
@@ -4,6 +4,7 @@ import com.cleanroommc.modularui.core.mixin.GuiContainerAccessor;
 import com.cleanroommc.modularui.screen.ClientScreenHandler;
 import com.cleanroommc.modularui.screen.ModularScreen;
 
+import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.inventory.Slot;
@@ -82,7 +83,7 @@ public interface IMuiScreen {
     }
 
     /**
-     * Hovering widget is handled by {@link com.cleanroommc.modularui.screen.viewport.GuiContext}.
+     * Hovering widget is handled by {@link ModularGuiContext}.
      * If it detects a slot, this method is called. Only affects {@link GuiContainer GuiContainers}.
      *
      * @param slot hovered slot
