+++ src/main/java/com/cleanroommc/modularui/api/widget/IGuiElement.java
@@ -3,7 +3,7 @@ package com.cleanroommc.modularui.api.widget;
 import com.cleanroommc.modularui.api.layout.IResizeable;
 import com.cleanroommc.modularui.screen.ModularPanel;
 import com.cleanroommc.modularui.screen.ModularScreen;
-import com.cleanroommc.modularui.screen.viewport.GuiContext;
+import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import com.cleanroommc.modularui.widget.sizer.Area;
 
 /**
@@ -49,7 +49,7 @@ public interface IGuiElement {
      *
      * @param context gui context
      */
-    void draw(GuiContext context);
+    void draw(ModularGuiContext context);
 
     /**
      * Called when the mouse enters the area of this element
