+++ src/main/java/com/cleanroommc/modularui/drawable/Rectangle.java
@@ -1,6 +1,5 @@
 package com.cleanroommc.modularui.drawable;
 
-import com.cleanroommc.modularui.api.ITheme;
 import com.cleanroommc.modularui.api.drawable.IDrawable;
 import com.cleanroommc.modularui.screen.viewport.GuiContext;
 import com.cleanroommc.modularui.theme.WidgetTheme;
