+++ src/main/java/com/cleanroommc/modularui/drawable/Stencil.java
@@ -13,6 +13,8 @@ import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import org.jetbrains.annotations.Nullable;
 import org.lwjgl.opengl.GL11;
 
+import java.awt.Rectangle;
+
 public class Stencil {
 
     // Stores a stack of areas that are transformed, so it represents the actual area
@@ -37,15 +39,15 @@ public class Stencil {
         GL11.glStencilMask(0x00);
     }
 
-    public static void apply(Area area, @Nullable GuiContext context) {
+    public static void apply(Rectangle area, @Nullable GuiContext context) {
         apply(area.x, area.y, area.width, area.height, context);
     }
 
-    public static void applyAtZero(Area area, @Nullable GuiContext context) {
+    public static void applyAtZero(Rectangle area, @Nullable GuiContext context) {
         apply(0, 0, area.width, area.height, context);
     }
 
-    public static void applyTransformed(Area area) {
+    public static void applyTransformed(Rectangle area) {
         applyTransformed(area.x, area.y, area.width, area.height);
     }
