+++ src/main/java/com/cleanroommc/modularui/widget/DraggableWidget.java
@@ -3,7 +3,7 @@ package com.cleanroommc.modularui.widget;
 import com.cleanroommc.modularui.api.layout.IViewport;
 import com.cleanroommc.modularui.api.layout.IViewportStack;
 import com.cleanroommc.modularui.api.widget.IDraggable;
-import com.cleanroommc.modularui.screen.viewport.GuiContext;
+import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import com.cleanroommc.modularui.utils.HoveredWidgetList;
 import com.cleanroommc.modularui.widget.sizer.Area;
 
@@ -25,7 +25,7 @@ public class DraggableWidget<W extends DraggableWidget<W>> extends Widget<W> imp
     }
 
     @Override
-    public void drawMovingState(GuiContext context, float partialTicks) {
+    public void drawMovingState(ModularGuiContext context, float partialTicks) {
         WidgetTree.drawTree(this, context, true);
     }
