+++ src/main/java/com/cleanroommc/modularui/widgets/CategoryList.java
@@ -5,7 +5,7 @@ import com.cleanroommc.modularui.api.layout.ILayoutWidget;
 import com.cleanroommc.modularui.api.widget.IWidget;
 import com.cleanroommc.modularui.api.widget.Interactable;
 import com.cleanroommc.modularui.drawable.GuiTextures;
-import com.cleanroommc.modularui.screen.viewport.GuiContext;
+import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import com.cleanroommc.modularui.theme.WidgetTheme;
 import com.cleanroommc.modularui.utils.Alignment;
 import com.cleanroommc.modularui.widget.ParentWidget;
@@ -25,7 +25,7 @@ public class CategoryList extends ParentWidget<CategoryList> implements Interact
     private IDrawable collapsedOverlay;
 
     @Override
-    public void drawOverlay(GuiContext context, WidgetTheme widgetTheme) {
+    public void drawOverlay(ModularGuiContext context, WidgetTheme widgetTheme) {
         super.drawOverlay(context, widgetTheme);
         if (this.expanded) {
             this.expandedOverlay.drawAtZero(context, getArea(), widgetTheme);
