+++ src/main/java/com/cleanroommc/modularui/widgets/ScrollingTextWidget.java
@@ -2,7 +2,7 @@ package com.cleanroommc.modularui.widgets;
 
 import com.cleanroommc.modularui.api.drawable.IKey;
 import com.cleanroommc.modularui.drawable.text.TextRenderer;
-import com.cleanroommc.modularui.screen.viewport.GuiContext;
+import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import com.cleanroommc.modularui.theme.WidgetTheme;
 import com.cleanroommc.modularui.utils.Alignment;
 
@@ -54,7 +54,7 @@ public class ScrollingTextWidget extends TextWidget {
     }
 
     @Override
-    public void draw(GuiContext context, WidgetTheme widgetTheme) {
+    public void draw(ModularGuiContext context, WidgetTheme widgetTheme) {
         checkString();
         TextRenderer renderer = TextRenderer.SHARED;
         renderer.setColor(getColor());
