+++ src/main/java/com/cleanroommc/modularui/widgets/TextWidget.java
@@ -2,7 +2,7 @@ package com.cleanroommc.modularui.widgets;
 
 import com.cleanroommc.modularui.api.drawable.IKey;
 import com.cleanroommc.modularui.drawable.text.TextRenderer;
-import com.cleanroommc.modularui.screen.viewport.GuiContext;
+import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import com.cleanroommc.modularui.theme.WidgetTheme;
 import com.cleanroommc.modularui.utils.Alignment;
 import com.cleanroommc.modularui.widget.Widget;
@@ -23,7 +23,7 @@ public class TextWidget extends Widget<TextWidget> {
     }
 
     @Override
-    public void draw(GuiContext context, WidgetTheme widgetTheme) {
+    public void draw(ModularGuiContext context, WidgetTheme widgetTheme) {
         TextRenderer renderer = TextRenderer.SHARED;
         renderer.setColor(this.color != null ? this.color : widgetTheme.getTextColor());
         renderer.setAlignment(this.alignment, getArea().w() + this.scale, getArea().h());
